e2e: valSig & disabled staking checks
